Changed Stats to use integers. Set enemy stats to load in map file.

+++ core/src/com/yosefu/game/Stats.java
@@ -1,12 +1,12 @@
 package com.yosefu.game;
 
 public class Stats {
-    public float health = 10;
-    public float maxHealth = 10;
-    public float pp = 10;
-    public float atk = 10;
+    public int health = 10;
+    public int maxHealth = 10;
+    public int pp = 10;
+    public int atk = 10;
 
-    public Stats(float health, float pp, float atk) {
+    public Stats(int health, int pp, int atk) {
         this.health = health;
         maxHealth = health;
         this.pp = pp;
